Tests/LibWeb/Text/expected/Wasm/WebAssembly-memory-reexport.txt

Re-exported memory equals imported memory: true
